FEAT: Add belongings list
