+++ src/components/layouts/PlanView.tsx
@@ -1,17 +1,14 @@
 import * as React from 'react'
 import Box from '@mui/material/Box'
-import IconButton from '@mui/material/IconButton'
 import Stack from '@mui/material/Stack'
-import SvgIcon from '@mui/material/SvgIcon'
 import Tab from '@mui/material/Tab'
 import Tabs from '@mui/material/Tabs'
-import Typography from '@mui/material/Typography'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { faEdit } from '@fortawesome/free-solid-svg-icons'
 import Image from 'next/image'
 
 import { useTravelPlan } from 'hooks/useTravelPlan'
-import SpotsList from 'components/modules/SpotsList'
+import TabPanel from 'components/modules/TabPanel'
+import PlanInfo from './PlanInfo'
+import BelongingsList from './BelongingsList'
 
 const PlanView = () => {
   const [plan] = useTravelPlan()
@@ -40,41 +37,12 @@ const PlanView = () => {
           <Tab label="持ち物" />
         </Tabs>
       </Box>
-      <Box p={2}>
-        <Stack spacing={4}>
-          <Stack
-            direction="row"
-            justifyContent="space-between"
-            alignItems="start">
-            <Stack>
-              <Typography variant="h1" noWrap>
-                {plan?.title}
-              </Typography>
-              <Typography variant="subtitle2">
-                {`${plan.start.toLocaleDateString()} ~ ${plan.end.toLocaleDateString()}`}
-              </Typography>
-              <Typography variant="caption">
-                {`from ${plan.home.name} to ${plan.destination.name}`}
-              </Typography>
-            </Stack>
-            <IconButton disabled>
-              <SvgIcon>
-                <FontAwesomeIcon icon={faEdit} />
-              </SvgIcon>
-            </IconButton>
-          </Stack>
-          <Stack spacing={2}>
-            <Typography variant="h2">行きたいところリスト</Typography>
-            {plan.waypoints.length > 0 ? (
-              <SpotsList spots={plan.waypoints} />
-            ) : (
-              <Typography variant="subtitle1">
-                地図上で行きたい場所を選んでください。
-              </Typography>
-            )}
-          </Stack>
-        </Stack>
-      </Box>
+      <TabPanel value={value} index={0} p={2}>
+        <PlanInfo plan={plan} />
+      </TabPanel>
+      <TabPanel value={value} index={1} p={2}>
+        <BelongingsList />
+      </TabPanel>
     </Stack>
   )
 }
