+++ src/components/modules/TabPanel.tsx
@@ -1,21 +1,20 @@
 import * as React from 'react'
-import Box from '@mui/material/Box'
+import Box, { BoxProps } from '@mui/material/Box'
 
-type Props = {
+type Props = BoxProps & {
   index: number
   value: number
 }
-const TabPanel: React.FC<Props> = ({ children, value, index }) => {
+const TabPanel: React.FC<Props> = ({ children, value, index, ...props }) => {
   return (
     <Box
       width="100%"
       height="100%"
       role="tabpanel"
-      position="absolute"
-      overflow="hidden auto"
       hidden={value !== index}
       id={`full-width-tabpanel-${index}`}
-      aria-labelledby={`full-width-tab-${index}`}>
+      aria-labelledby={`full-width-tab-${index}`}
+      {...props}>
       {value === index && children}
     </Box>
   )
